bifurcate new and old one click login links
