+++ apps/passport-client/components/screens/LoginScreens/OneClickLoginScreen.tsx
@@ -1,34 +1,22 @@
-import { requestGenericIssuanceTicketPreviews } from "@pcd/passport-interface";
-import { Button, Spacer } from "@pcd/passport-ui";
-import { IPODTicketData } from "@pcd/pod-ticket-pcd";
-import { PODTicketCardBodyImpl } from "@pcd/pod-ticket-pcd-ui";
-import { useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect } from "react";
 import { useParams } from "react-router-dom";
-import { appConfig } from "../../../src/appConfig";
 import { useDispatch, useSelf } from "../../../src/appHooks";
 import { MaybeModal } from "../../modals/Modal";
 import { AppContainer } from "../../shared/AppContainer";
-import {
-  CardBodyContainer,
-  CardHeader,
-  CardOutlineExpanded
-} from "../../shared/PCDCard";
 import { ScreenLoader } from "../../shared/ScreenLoader";
 
 /**
- * format: http://localhost:3000/#/one-click-login/:email/:code/:targetFolder/:pipelineId?/:serverUrl?
+ * format: http://localhost:3000/#/one-click-login/:email/:code/:targetFolder
  * - `code` is the pretix or lemonade order code
  * - `email` is the email address of the ticket to whom the ticket was issued
  * - `targetFolder` is the folder to redirect to after login. optional.
  * example: http://localhost:3000/#/one-click-login/ivan@0xparc.org/123456/0xPARC%2520Summer%2520'24
  */
 export function OneClickLoginScreen(): JSX.Element | null {
-  const self = useSelf();
   const dispatch = useDispatch();
-  const { email, code, targetFolder, pipelineId, serverUrl } = useParams();
-  const [ticketPreviews, setTicketPreviews] = useState<IPODTicketData[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | undefined>();
+  const { email, code, targetFolder } = useParams();
+
+  const self = useSelf();
 
   const redirectToTargetFolder = useCallback(() => {
     if (targetFolder) {
@@ -38,42 +26,11 @@ export function OneClickLoginScreen(): JSX.Element | null {
     }
   }, [targetFolder]);
 
-  const handleLoadTicketPreviews = useCallback(async () => {
-    if (!email || !code) {
-      return;
-    }
-    try {
-      const previewRes = await requestGenericIssuanceTicketPreviews(
-        serverUrl ?? appConfig.zupassServer,
-        email,
-        code,
-        pipelineId
-      );
-
-      setLoading(false);
-      if (previewRes.success) {
-        setTicketPreviews(previewRes.value.tickets);
-      } else {
-        setError(previewRes.error);
-      }
-    } catch (err) {
-      await dispatch({
-        type: "error",
-        error: {
-          title: "An error occured",
-          message: (err as Error).message || "An error occured"
-        }
-      });
-    }
-  }, [email, code, serverUrl, pipelineId, dispatch]);
-
   const handleOneClickLogin = useCallback(async () => {
     if (!email || !code) {
       return;
     }
-
     try {
-      setLoading(true);
       await dispatch({
         type: "one-click-login",
         email,
@@ -81,7 +38,6 @@ export function OneClickLoginScreen(): JSX.Element | null {
         targetFolder
       });
     } catch (err) {
-      setLoading(false);
       await dispatch({
         type: "error",
         error: {
@@ -90,7 +46,7 @@ export function OneClickLoginScreen(): JSX.Element | null {
         }
       });
     }
-  }, [email, code, dispatch, targetFolder]);
+  }, [dispatch, email, code, targetFolder]);
 
   useEffect(() => {
     if (process.env.ONE_CLICK_LOGIN_ENABLED !== "true") {
@@ -115,12 +71,12 @@ export function OneClickLoginScreen(): JSX.Element | null {
     } else if (!email || !code) {
       window.location.hash = "#/";
     } else {
-      handleLoadTicketPreviews();
+      handleOneClickLogin();
     }
   }, [
     self,
     targetFolder,
-    handleLoadTicketPreviews,
+    handleOneClickLogin,
     redirectToTargetFolder,
     email,
     code
@@ -130,39 +86,7 @@ export function OneClickLoginScreen(): JSX.Element | null {
     <>
       <MaybeModal fullScreen />
       <AppContainer bg="primary">
-        {loading && <ScreenLoader />}
-
-        {error && <div>{error}</div>}
-
-        {!loading && (
-          <>
-            <Spacer h={32} />
-
-            {ticketPreviews.length === 0 ? (
-              <div>No tickets found</div>
-            ) : (
-              <>
-                {ticketPreviews.map((ticket, i) => (
-                  <CardOutlineExpanded key={i}>
-                    <CardBodyContainer>
-                      <CardHeader isMainIdentity={true}>
-                        {ticket.eventName} ({ticket.ticketName})
-                      </CardHeader>
-                      <PODTicketCardBodyImpl
-                        idBasedVerifyURL=""
-                        ticketData={ticket}
-                        key={ticket.ticketId}
-                      />
-                    </CardBodyContainer>
-                  </CardOutlineExpanded>
-                ))}
-                <Spacer h={16} />
-              </>
-            )}
-
-            <Button onClick={handleOneClickLogin}>Continue to Zupass</Button>
-          </>
-        )}
+        <ScreenLoader />
       </AppContainer>
     </>
   );
