+++ apps/passport-client/pages/index.tsx
@@ -34,6 +34,7 @@ import { AlreadyRegisteredScreen } from "../components/screens/LoginScreens/Alre
 import { CreatePasswordScreen } from "../components/screens/LoginScreens/CreatePasswordScreen";
 import { LoginInterstitialScreen } from "../components/screens/LoginScreens/LoginInterstitialScreen";
 import { LoginScreen } from "../components/screens/LoginScreens/LoginScreen";
+import { NewOneClickLoginScreen } from "../components/screens/LoginScreens/NewOneClickLoginScreen";
 import { NewPassportScreen } from "../components/screens/LoginScreens/NewPassportScreen";
 import { OneClickLoginScreen } from "../components/screens/LoginScreens/OneClickLoginScreen";
 import { PrivacyNoticeScreen } from "../components/screens/LoginScreens/PrivacyNoticeScreen";
@@ -147,8 +148,12 @@ function RouterImpl(): JSX.Element {
           <Route path="add-email" element={<AddEmailScreen />} />
           <Route path="remove-email" element={<RemoveEmailScreen />} />
           <Route
-            path="one-click-login/:email/:code/:targetFolder/:pipelineId?/:serverUrl?"
+            path="one-click-login/:email/:code/:targetFolder"
             element={<OneClickLoginScreen />}
+          />
+          <Route
+            path="one-click-preview/:email/:code/:targetFolder/:pipelineId?/:serverUrl?"
+            element={<NewOneClickLoginScreen />}
           />
           <Route
             path="enter-confirmation-code"
